✨ feat(auth): api for sign-up

+++ src/auth/authentication.service.ts
@@ -5,10 +5,15 @@ import { CookieOptions, Response } from 'express'
 import { SignUpReq } from './dto/request/signup.req'
 import { EmailVO } from '../users/vo/email.vo'
 import { InputPasswordVO } from '../users/vo/password.vo'
+import { CompanyService } from '../company/company.service'
 
 @Injectable()
 export class AuthenticationService {
-  constructor(private usersService: UsersService, private jwtService: JwtService) {}
+  constructor(
+    private readonly jwtService: JwtService,
+    private readonly usersService: UsersService,
+    private readonly companyService: CompanyService,
+  ) {}
 
   async signIn(email: EmailVO, password: InputPasswordVO, response: Response) {
     const user = await this.usersService.findUserIdByEmail(email)
@@ -35,13 +40,20 @@ export class AuthenticationService {
     const { code, password, ...rest } = signUpReq
 
     const invitationMail = await this.usersService.findInvitationMail(code)
-    // if (!invitationMail) throw new NotFoundException()
+    if (!invitationMail) throw new NotFoundException('Invitation Not Found')
 
-    // Check Company
+    const company = await this.companyService.findCompanyById(invitationMail.companyId)
+    if (!company) throw new NotFoundException('Company Not Found')
+
+    const user = await this.usersService.insertUser(rest, new InputPasswordVO(password))
 
     // Give User Role
+    await this.companyService.giveUserRole({
+      userId: user.id,
+      role: invitationMail.role,
+      companyType: invitationMail.companyType,
+    })
 
-    await this.usersService.insertUser(rest, new InputPasswordVO(password))
     await this.usersService.deleteInvitationMail(code)
   }
